src/ubuntu/dotfiles/helpers/decorators/installation_step.py
"""Defines a decorator to log the progress of an installation step."""

from functools import wraps
from logging import exception as logException
from logging import info as logInfo
from typing import Any, Callable

from dotfiles.type_definitions import InstallationStepReturnValueT


def installationStep(
    function: Callable[..., InstallationStepReturnValueT]
) -> Callable[..., InstallationStepReturnValueT]:
    """Log the progress (start and completion/failure) of an installation step."""

    @wraps(function)
    def runInstallationStep(*args: Any, **kwargs: Any) -> InstallationStepReturnValueT:
        stepName = function.__name__
        logInfo("--- Starting step: `%s` ---", stepName)

        try:
            returnValue = function(*args, **kwargs)
        except Exception as exception:
            logException("--- Failed step: `%s` ---\n", stepName)
            raise exception

        logInfo("--- Finished step: `%s` ---\n", stepName)
        return returnValue

    return runInstallationStep
